+++ src/hooks/login/useWalletConnectV2Login.ts
@@ -15,15 +15,13 @@ import {
   walletConnectV2OptionsSelector
 } from 'reduxStore/selectors/networkConfigSelectors';
 import { setWalletConnectLogin } from 'reduxStore/slices';
-import {
-  LoginMethodsEnum,
-  DappCoreWCV2CustomMethodsEnum
-} from 'types/enums.types';
+import { LoginMethodsEnum } from 'types/enums.types';
 import { getIsProviderEqualTo } from 'utils/account/getIsProviderEqualTo';
 import { getIsLoggedIn } from 'utils/getIsLoggedIn';
 import { optionalRedirect } from 'utils/internal';
 import { logout } from 'utils/logout';
 import {
+  WalletConnectOptionalMethodsEnum,
   WalletConnectV2Provider,
   SessionEventTypes,
   PairingTypes
@@ -95,8 +93,14 @@ export const useWalletConnectV2Login = ({
   const isInitialisingRef = useRef<boolean>(false);
 
   const dappMethods: string[] = [
-    DappCoreWCV2CustomMethodsEnum.mvx_cancelAction
+    WalletConnectOptionalMethodsEnum.CANCEL_ACTION
   ];
+  if (tokenToSign) {
+    dappMethods.push(WalletConnectOptionalMethodsEnum.SIGN_LOGIN_TOKEN);
+  }
+  if (hasNativeAuth) {
+    dappMethods.push(WalletConnectOptionalMethodsEnum.SIGN_NATIVE_AUTH_TOKEN);
+  }
 
   const uriDeepLink = !isLoading
     ? `${walletConnectDeepLink}?wallet-connect=${encodeURIComponent(wcUri)}`
@@ -153,13 +157,13 @@ export const useWalletConnectV2Login = ({
         return;
       }
 
-      const address = await provider.getAddress();
+      const address = await providerRef.current?.getAddress();
       if (!address) {
         console.warn('Login cancelled.');
         return;
       }
 
-      const signature = await provider.getSignature();
+      const signature = await providerRef.current?.getSignature();
       const loginActionData = {
         address: address,
         loginMethod: LoginMethodsEnum.walletconnectv2
@@ -199,13 +203,27 @@ export const useWalletConnectV2Login = ({
       return;
     }
 
+    isInitialisingRef.current = true;
+
+    if (providerRef.current?.walletConnector) {
+      providerRef.current.init();
+
+      setAccountProvider(providerRef.current);
+      isInitialisingRef.current = false;
+      canLoginRef.current = true;
+
+      if (loginProvider) {
+        generateWcUri();
+      }
+
+      return;
+    }
+
     const providerHandlers = {
       onClientLogin: handleOnLogin,
       onClientLogout: handleOnLogout,
       onClientEvent: handleOnEvent
     };
-
-    isInitialisingRef.current = true;
     const newProvider = new WalletConnectV2Provider(
       providerHandlers,
       chainId,
@@ -213,16 +231,19 @@ export const useWalletConnectV2Login = ({
       walletConnectV2ProjectId,
       walletConnectV2Options
     );
-
     await newProvider.init();
+
+    setAccountProvider(newProvider);
+    providerRef.current = newProvider;
     isInitialisingRef.current = false;
     canLoginRef.current = true;
-    setAccountProvider(newProvider);
-    setWcPairings(newProvider.pairings);
-    providerRef.current = newProvider;
+
     if (loginProvider) {
+      setWcPairings(newProvider.pairings);
       generateWcUri();
     }
+
+    return;
   }
 
   async function connectExisting(pairing: PairingTypes.Struct) {