+++ src/reduxStore/slices/toastsSlice.ts
@@ -1,25 +1,18 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { REHYDRATE } from 'redux-persist';
-import { TRANSACTION_STATUS_TOAST_ID } from 'constants/transactionStatus';
 import { ToastsEnum } from 'types';
+import { CustomToastType, TransactionToastType } from 'types/toasts.types';
 import { getUnixTimestamp } from 'utils/dateTime/getUnixTimestamp';
-import {
-  CustomToastType,
-  FailTransactionToastType,
-  TransactionToastType
-} from '../../types/toasts.types';
 import { logoutAction } from '../commonActions';
 
 export interface ToastsSliceState {
   customToasts: CustomToastType[];
   transactionToasts: TransactionToastType[];
-  failTransactionToast: FailTransactionToastType | null;
 }
 
 const initialState: ToastsSliceState = {
   customToasts: [],
-  transactionToasts: [],
-  failTransactionToast: null
+  transactionToasts: []
 };
 
 export const toastsSlice = createSlice({
@@ -66,18 +59,6 @@ export const toastsSlice = createSlice({
       state.transactionToasts = state.transactionToasts.filter((toast) => {
         return toast.toastId !== action.payload;
       });
-    },
-    addFailTransactionToast: (
-      state: ToastsSliceState,
-      action: PayloadAction<FailTransactionToastType>
-    ) => {
-      state.failTransactionToast = {
-        ...action.payload,
-        toastId: TRANSACTION_STATUS_TOAST_ID
-      };
-    },
-    removeFailTransactionToast: (state: ToastsSliceState) => {
-      state.failTransactionToast = null;
     }
   },
 
@@ -99,9 +80,7 @@ export const {
   addCustomToast,
   removeCustomToast,
   addTransactionToast,
-  removeTransactionToast,
-  addFailTransactionToast,
-  removeFailTransactionToast
+  removeTransactionToast
 } = toastsSlice.actions;
 
 export default toastsSlice.reducer;