+++ src/utils/walletconnect/__sdkWalletconnectProvider.ts
@@ -3,16 +3,20 @@ import {
   PairingTypes,
   EngineTypes
 } from '@multiversx/sdk-wallet-connect-provider';
-import { WalletConnectProvider } from '@multiversx/sdk-wallet-connect-provider/out/walletConnectProvider';
+import {
+  Operation as WalletConnectMethodsEnum,
+  OptionalOperation as WalletConnectOptionalMethodsEnum
+} from '@multiversx/sdk-wallet-connect-provider/out/operation';
 import { WalletConnectV2Provider } from '@multiversx/sdk-wallet-connect-provider/out/walletConnectV2Provider';
 
 /**
  * These members are ingnored when compiling to commonJS
  */
 export {
-  SessionEventTypes,
-  PairingTypes,
-  WalletConnectV2Provider,
   EngineTypes,
-  WalletConnectProvider
+  PairingTypes,
+  SessionEventTypes,
+  WalletConnectMethodsEnum,
+  WalletConnectOptionalMethodsEnum,
+  WalletConnectV2Provider
 };